Re-org per review comments.
